requirements.txt
tensorflow~=2.9.1
opencv-python~=4.6.0.66
numpy~=1.20.3
pillow~=8.4.0
streamlit~=1.10.0
matplotlib~=3.4.3
keras~=2.9.0
scikit-learn~=1.0.2